Ghost Jump scare
Changed flash light code to adapt singleton pattern.
Finish Ghost appear JS in scene and in code, able to generate sound and flashy light effects.
Add in a FlashyLightHelper as an attachable component anywhere you can use. They're customizable, just drag in the light reference, or even in the future an array of lights, can change the light color, can play with the flash count, and will support random flash interval in the near future.